contracts: update precompile wrappers

+++ contracts/contracts/WrappedROSE.sol
@@ -5,9 +5,10 @@ import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 import "@openzeppelin/contracts/token/ERC20/extensions/ERC20Burnable.sol";
 
 contract WrappedROSE is ERC20, ERC20Burnable {
-    event  Deposit(address indexed dst, uint wad);
-    event  Withdrawal(address indexed src, uint wad);
+    event Deposit(address indexed dst, uint256 wad);
+    event Withdrawal(address indexed src, uint256 wad);
 
+    //solhint-disable-next-line no-empty-blocks
     constructor() ERC20("Wrapped ROSE", "wROSE") {}
 
     function deposit() external payable {
